latest ecom-app

+++ server/App/controllers/admin/categoryController.js
@@ -1,5 +1,10 @@
 const { categoryModel } = require("../../models/categoryModel")
-let fs = require("fs")
+let fs = require("fs");
+const { sub_subcategoryModel } = require("../../models/sub-subcategoryModel");
+const { subcategoryModel } = require("../../models/subCategoryModel");
+//const {default: slugify} =require("slugify")
+const {default: slugify} = require("slugify")
+
 let categoryInsert = async (req, res) => {
 
     let { categoryName, categoryOrder } = req.body
@@ -7,7 +12,9 @@ let categoryInsert = async (req, res) => {
     let obj = {
         categoryName,
         categoryOrder,
-        categoryStatus
+        categoryStatus,
+      /*   slug:slugify(categoryName, {lower:true})
+ */
     }
     if (req.file) {
         if (req.file.filename) {
@@ -79,7 +86,7 @@ let categoryView = async (req, res) => {
         }
 
         res.send(obj)
-         console.log("suc", obj)
+        console.log("suc", obj)
     }
     catch (error) {
         obj = {
@@ -93,23 +100,102 @@ let categoryView = async (req, res) => {
 
 let categoryDelete = async (req, res) => {
     let { ids } = req.body
-    let categoryView = await categoryModel.find({ _id: ids }).select("categoryImage")
+    let obj;
+    try {
+
+        // step1: delete sub subcategory pics
+        let sub_subcatpics = await sub_subcategoryModel.find({ parentCategory: ids }).select("sub_subcatImage")
 
 
-    for (let v of categoryView) {
-        let deletePath = "uploads/category/" + v.categoryImage;
-        fs.unlinkSync(deletePath)
+        for (let v of sub_subcatpics) {
+            let deletePath = "uploads/sub_subcategory/" + v.sub_subcatImage;
+            fs.unlinkSync(deletePath)
+
+        }
+        // delete sub sub category
+        const sub_subcategoryDelete = await sub_subcategoryModel.deleteMany({ parentCategory: ids });
+
+        // step: 2 if sub subcategory delete then delete subcategory
+
+        if (sub_subcategoryDelete.acknowledged && sub_subcategoryDelete.deletedCount >= 0) {
+            // delete sub category pics 
+            let subcategorypics = await subcategoryModel.find({ parentCategory: ids }).select("subcategoryImage")
+
+
+            for (let v of subcategorypics) {
+                let deletePath = "uploads/subcategory/" + v.subcategoryImage;
+                fs.unlinkSync(deletePath)
+
+            }
+
+            // delete sub category
+            let subcategoryDelete = await subcategoryModel.deleteMany({ parentCategory: ids });
+
+            // step:3 if sub category also deleted then delete parent category
+            if (subcategoryDelete.acknowledged && subcategoryDelete.deletedCount >= 0) {
+
+                //delete photos of parent category
+                let categorypics = await categoryModel.find({ _id: ids }).select("categoryImage")
+
+
+                for (let v of categorypics) {
+                    let deletePath = "uploads/category/" + v.categoryImage;
+                    fs.unlinkSync(deletePath)
+
+                }
+
+                //delete parent category
+                let parentcatDelete = await categoryModel.deleteMany({ _id: ids })
+                if (parentcatDelete.deletedCount === 1) {
+
+                    obj = {
+                        status: 1,
+                        msg: "parent category deleted successfully!!"
+                    }
+                    res.send(obj);
+                    console.log(obj)
+                }
+                else {
+
+                    obj = {
+                        status: 0,
+                        msg: "parentCategory not found"
+                    }
+                    res.send(obj);
+                    console.log(obj)
+                }
+
+            }
+            else {
+                obj = {
+                    status: 0,
+                    msg: "Failed to delete  subcategories"
+                }
+                res.send(obj);
+                console.log(obj);
+            }
+        }
+        else {
+
+            obj = {
+                status: 0,
+                msg: "Failed to delete sub subcategories"
+            }
+            res.send(obj);
+        }
+
 
     }
+    catch (error) {
 
-    let data = await categoryModel.deleteMany({ _id: ids })
-
-    let obj = {
-        status: 1,
-        msg: "Category is deleted successfully!!"
+        console.log(error);
+        obj = {
+            status: 0,
+            msg: error + " An error occurred while deleting categories"
+        }
+        res.send(obj);
     }
 
-    res.send(obj)
 }
 
 let changeStatus = async (req, res) => {
@@ -134,8 +220,8 @@ let categoryUpdate = async (req, res) => {
 
     let categoryAdd = {
         categoryName,
-        categoryOrder,
-        categoryStatus: true
+        categoryOrder
+
     }
 
     if (req.file) {
@@ -143,7 +229,7 @@ let categoryUpdate = async (req, res) => {
             categoryAdd['categoryImage'] = req.file.filename
         }
     }
-   
+
     try {
         let categoryView = await categoryModel.find({ _id: id }).select("categoryImage")
 
@@ -178,7 +264,7 @@ let singleCategoryView = async (req, res) => {
         obj = {
             status: 1,
             msg: "category view ",
-             staticPath: process.env.CATEGORYIMAGEPATH,
+            staticPath: process.env.CATEGORYIMAGEPATH,
             data
         }
         res.send(obj)
@@ -192,4 +278,4 @@ let singleCategoryView = async (req, res) => {
     }
 }
 
-module.exports = { categoryInsert, categoryView,categoryUpdate, categoryDelete, changeStatus, singleCategoryView }
+module.exports = { categoryInsert, categoryView, categoryUpdate, categoryDelete, changeStatus, singleCategoryView }
